requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-jwt-extended==4.3.1
python-dotenv==0.19.0
requests==2.26.0
pytest==6.2.5
black==21.9b0
flake8==3.9.2
